Add more monitoring pages and other changes
