+++ app/models/campaign_message/prior_gyr.rb
@@ -6,6 +6,10 @@ module CampaignMessage
       'campaign_messages.prior_gyr'.freeze
     end
 
+    def self.batch_scope
+      :prior_gyr
+    end
+
     def email_subject(contact:, **args)
       I18n.t("campaign_messages.prior_gyr.email.subject", **vars(contact), **args)
     end
